--- webui/react/src/hooks/useModal/Columns/useModalColumnsCustomize.module.scss
@@ -1,60 +1,0 @@
-.base {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-
-  .columns {
-    display: flex;
-    flex-direction: row;
-    gap: 6px;
-
-    .column {
-      flex-basis: 50%;
-
-      .visibleTitleRow {
-        display: flex;
-        justify-content: space-between;
-
-        & > * {
-          align-self: start;
-          border: 0;
-          height: 100%;
-          padding: 0;
-        }
-      }
-      .listContainer {
-        border: solid var(--theme-stroke-width) var(--theme-surface-border);
-        border-radius: var(--theme-border-radius);
-
-        ul {
-          height: 200px;
-          list-style-type: none;
-          margin-bottom: 0;
-          overflow-y: auto;
-          padding-left: 0;
-          padding-top: 6px;
-
-          li {
-            cursor: pointer;
-            padding-left: 6px;
-
-            &:hover {
-              background-color: var(--theme-surface-strong);
-            }
-            mark {
-              background-color: #faed27;
-              padding: 0;
-            }
-          }
-        }
-      }
-      button {
-        padding: 6px;
-      }
-    }
-  }
-  h2 {
-    color: var(--theme-colors-monochrome-5);
-    font-size: 14px;
-  }
-}
